CA2(176-177-206-219)/Rasoi-Genie/backend/requirements.txt
# FastAPI and related
fastapi==0.95.2
uvicorn==0.24.0

# Database
sqlalchemy==2.0.23
pymysql==1.1.0
mysql-connector-python==8.3.0

# Auth and utils
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==1.10.12       
email-validator==1.3.1 

# LangChain and AI
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
openai==1.6.1
